requirements.txt
Pillow~=8.3.1
PyYAML~=5.4.1
easyocr~=1.4
flake8~=3.9.2
jupyter~=1.0.0
matplotlib~=3.4.2
nltk~=3.6.2
numpy~=1.21.0
paddleocr~=2.3.0.1
pandas~=1.3.0
pyenchant~=3.2.0
pytesseract~=0.3.7
pytorch-lightning~=1.4.7
opencv-python>=4.5.2.54
scikit-image~=0.17.2
scikit-learn~=0.24.2
scipy~=1.7.0
tqdm~=4.61.2

# torch>=1.8.1
# torchvision>=0.9.1
#torch>=1.8.1+cu111
#torchvision>=0.9.1+cu111
#nlp>=0.4.0
#transformers>=4.10.2
